Add file_crlf to OpamParserTypes.opamfile
The file_crlf field records whether the opam file was read from a file
where the line endings were CRLF-encoded. This changes the signatures
for the lexer, as state must now be threaded through each call. The EOF
token in OpamBaseParser has a bool option parameter added to it which
indicates if the file line-endings were CRLF-encoded (Some true),
LF-encoded (Some false), or mixed (None). If the file endings are mixed,
then OpamParser will default to assuming CRLF-encoding on Windows and
LF-encoding everywhere else, including Cygwin.