Escape any newline at the start of triple-quotes
The OpamPrinter.format_value formats a string which contains newlines
as:

"""
line 1
line 2"""

The problem before is that the newline after the opening triple quote is
not escaped, so the string represented gains a newline. It should in
fact be:

"""\
line 1
line 2"""

so that round-tripping works. This commit adds a further optimisation so
that strings which actually begin with a newline display as:

"""
line 1
line 2"""